requirements.txt
--index-url https://download.pytorch.org/whl/cu118

torch 
torchvision 
torchaudio
ipykernel
pyyaml
matplotlib
opencv-python
pandas
ultralytics
scikit-learn